backend/pyproject.toml
[project]
name = "work-log-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "atlassian-python-api>=4.0.7",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
